Refactor AICI to use WebAssembly Component Model
This is a significant change to the AICI Runtime (host) and AICI Controller
(guest) to use WASI components. As part of this change, a significant amount of
unsafe code is removed and the protocol is simplified to remove the need for
"BLOB" types and side-channels.

The protocol is documented in `wit/controller.wit`, and a minimal WASI runtime
is provided to AI CI controllers.

Some notes:
* The AICI runtime now longer directly reads and writes to the guest's memory.
  Instead, the guest provides a `Runner` resource (using WebAssembly Component
  terminology), which exposes the low-level protocol to the host as a
  constructor and trait with methods.
* The Blob protocols are removed entirely, replaced by the `Runner` resource.
  This and other side-channels for communicating with the runtime, e.g. allowed
  tokens (logit biases) outside of `MidProcessResult`, are removed.
* The (Variable) Storage and Tokenizer protocols are separate WebAssembly
  Components, which can be versioned independently of the runtime.
* Types are changed to be consistent with the WebAssembly interface, e.g.:
  `SeqId` is used in far more places to avoid casts.

+++ aicirt/src/api.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use crate::HashMap;
 
-pub type ModuleInstId = usize;
+pub type ModuleInstId = u64;
 
 #[derive(Serialize, Deserialize)]
 pub struct AiciPostPreProcessReq {
